requirements.txt
finch-api==1.7.0
Flask==3.0.3
Flask-Session==0.8.0
Flask-WTF==1.2.1
redis==5.1.1
requests==2.32.3
WTForms==3.1.2
